==> 8-Baseline-Experiments-with-other-Classifiers.ipynb <==
# Baseline Experiments with other Classifiers

For the paper, I performed additional experiments to analyze the default scores, obtained with the dummy classifier and SVMs.

## Preparing the dataset

Importing the necessary libraries

import json
import pandas as pd
#from copy import deepcopy
import re
#from tqdm import tqdm
import numpy as np
import sklearn.model_selection
from sklearn import metrics
from sklearn.metrics import classification_report, f1_score
from sklearn.dummy import DummyClassifier
from sklearn.pipeline import make_pipeline
from sklearn.naive_bayes import MultinomialNB,ComplementNB
from sklearn.linear_model import LogisticRegression
from sklearn.svm import SVC
from sklearn.ensemble import RandomForestClassifier
from sklearn.tree import DecisionTreeClassifier
from sklearn.metrics import ConfusionMatrixDisplay
import matplotlib.pyplot as plt
import sklearn.feature_extraction
from sklearn.svm import SVC


# Import the file with additional text representations (only the paragraphs marked to be kept
# in the original corpus are included)

with open("data/Language-Processed-GINCO.json") as f:
    dataset = json.load(f)

dataset[0]

dataset[0].keys()

### Downcasting number of labels

In these experiments, we will not use all of the texts but only texts from 5 main categories, meaning that some categories will be merged into them, whereas some categories with a very small frequency will be discarded. Additionally, the texts marked us hard, will be discarded (see notebook *1-Preparing_Data_Hyperparameter_Search*).

We will start with a reduced set of labels (primary_level_3), then merge News and Opinionated News, and discard some of the lables.

# merge News and Opinionated News
for i in dataset:
    if i["primary_level_3"] == "Opinionated News" or i["primary_level_3"] == "News/Reporting":
        i["primary_level_3"] = "News"

Let's create train:test:dev split that contains only the wanted labels.

downcasted_labels = ['Information/Explanation', 'Promotion', 'News', 'Forum', 'Opinion/Argumentation']

train = [i for i in dataset if i["split"] == "train" and i["primary_level_3"] in downcasted_labels and not i["hard"]]
test = [i for i in dataset if i["split"] == "test" and i["primary_level_3"] in downcasted_labels and not i["hard"]]
dev = [i for i in dataset if i["split"] == "dev" and i["primary_level_3"] in downcasted_labels and not i["hard"]]

print("The train-dev-test splits consist of the following numbers of examples:", len(train), len(test), len(dev))

print(f"Number of all texts is {len(train)+len(test)+len(dev)}")

# Create dataframes from the datasets
train_df = pd.DataFrame(train)

train_df.columns

test_df = pd.DataFrame(test)
dev_df = pd.DataFrame(dev)

print(f"test: {test_df.shape}, dev: {dev_df.shape}")

For the experiments, we'll use the baseline text ("baseline_text") and the "primary_level_3" labels.

train_df.columns

# Create X_train and Y_train parts, used for sci kit learning
# List of texts in training split
X_train = list(train_df.baseline_text)
# List of labels in training split
Y_train = list(train_df.primary_level_3)

# List of texts in test split
X_test = list(test_df.baseline_text)
# List of labels in test split
Y_test = list(test_df.primary_level_3)

print(len(X_train), len(Y_train), len(X_test), len(Y_test))

Y_test[:2]

# Create a list of labels
labels = list(train_df.primary_level_3.unique())
labels

# Create a TF-IDF representation of the text
def data_iterator(f):
    for token in f:
        yield token


def tokenizer(txt):
    """Simple whitespace tokenizer"""
    return txt.split()

iterator=data_iterator(X_train)
test_iterator=data_iterator(X_test)

vectorizer=sklearn.feature_extraction.text.TfidfVectorizer(tokenizer=tokenizer,use_idf=True,min_df=0.005)

d=vectorizer.fit_transform(iterator)

d_test=vectorizer.transform(test_iterator)

## Training with various Sci-Kit models

# Create a pipeline of models that you want to try:

pipelines=[]

"""
for model in [DummyClassifier(strategy="most_frequent"), DummyClassifier(strategy="stratified"), DecisionTreeClassifier(), MultinomialNB(), ComplementNB(), LogisticRegression(solver='saga'), SVC(C=0.5,kernel='linear',shrinking=False,probability=True),RandomForestClassifier()]:
    pipeline=make_pipeline(model)
    pipelines.append(pipeline)

"""

for model in [DummyClassifier(strategy="most_frequent"), DummyClassifier(strategy="stratified"), DecisionTreeClassifier(), MultinomialNB(), ComplementNB(), LogisticRegression(), SVC(),RandomForestClassifier()]:
    pipeline=make_pipeline(model)
    pipelines.append(pipeline)

pipelines

for i, pipeline in enumerate(pipelines):
    pipeline.fit(d, Y_train)

#Prediction from test dataset
model_name=[]
y_pred_list = []
micro_f1_array=[]
macro_f1_array = []
accuracy_array = []

print("Classifiation Report\n")
print("*****************************************************")
for i, pipeline in enumerate(pipelines):
    y_pred=pipeline.predict(d_test)
    y_pred_list.append(list(y_pred))
    print(pipelines[i].steps[0][0].upper())
    model_name.append(pipelines[i].steps[0][0].upper())

    micro_f1_array.append(round(f1_score(Y_test, y_pred, labels=labels, average ="micro"),3))
    macro_f1_array.append(round(f1_score(Y_test, y_pred, labels=labels, average ="macro"),3))
    accuracy_array.append(round(metrics.accuracy_score(Y_test, y_pred),3))
    print("\n",classification_report(Y_test, y_pred, zero_division = 0))
    print("*****************************************************")

results = {"model":model_name, "microF1": micro_f1_array, "macroF1":macro_f1_array, "accuracy":accuracy_array, "y_pred":y_pred_list}

results_df = pd.DataFrame(results)
results_df

# Append the results for the FastText model

ft = {"model": "FastText", "microF1": 0.56, "macroF1": 0.589}

results_df = results_df.append(ft, ignore_index = True)

results_df



# Show in markdown
print(results_df[["model","microF1", "macroF1"]].to_markdown(index = False))

# Save the results in a csv format
results_df.to_csv("results/additional_models_experiments.csv")
